% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmSparseNet.R
\docType{methods}
\name{cv.glmSparseNet.mclapply,matrix-method}
\alias{cv.glmSparseNet.mclapply,matrix-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{cv.glmSparseNet.mclapply}{matrix}(xdata, ydata, network,
  network.options = network.options.default(), ...)
}
\description{
Calculate GLM model with network-based regularization
}
