% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmSparseNet.R
\docType{methods}
\name{cv.glmSparseNet.mclapply,MultiAssayExperiment-method}
\alias{cv.glmSparseNet.mclapply,MultiAssayExperiment-method}
\title{Calculate GLM model with network-based regularization}
\usage{
\S4method{cv.glmSparseNet.mclapply}{MultiAssayExperiment}(xdata, ydata,
  network, experiment.name = NULL,
  network.options = network.options.default(), ...)
}
\arguments{
\item{experiment.name}{Name of experiment in MultiAssayExperiment}
}
\description{
Calculate GLM model with network-based regularization
}
