% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_index.R
\name{fit.risk}
\alias{fit.risk}
\title{Calculate c-index}
\usage{
fit.risk(coef.v, xdata, ydata, model.name, show.message = FALSE,
  n.cores = 1)
}
\arguments{
\item{n.cores}{}
}
\description{
Calculate c-index
}
