% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmsparsenet_custom.R
\name{glmSparseNet.cox_}
\alias{glmSparseNet.cox_}
\title{Call glmSparseNet model with Cox regression}
\usage{
glmSparseNet.cox_(xdata, ydata, target.vars, network = "correlation",
  xdata.digest = NULL, cache.prefix = "glmSparseNet.cache",
  force.recalc = FALSE, ...)
}
\arguments{
\item{xdata}{input matrix}

\item{ydata}{Surival data.dataframe with time and status columns}

\item{target.vars}{number of variables to target in calculations}

\item{network}{network to use, could be a matrix, a degree vector or a string, see ?glmSparseNet}

\item{xdata.digest}{sha256 checksum of xdata}

\item{cache.prefix}{prefix for the cache files}

\item{force.recalc}{force cache to be recalculated}
}
\value{
a glmnet model
}
\description{
This is an auxiliary method for the analysis. It uses the run.cache method
to cache the results and speed up analysis
}
