% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaplan_meier.R
\name{km.name}
\alias{km.name}
\title{Title}
\usage{
km.name(title, alpha, is.network = NULL)
}
\arguments{
\item{is.network}{}
}
\description{
Title
}
\examples{
km.name('This title', .6)
km.name('This title', 1)
km.name('This title', 0, 'glmnet')
km.name('This title', .7, 'orphan')
}
