% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmsparsenet_custom.R
\name{my.cv.glmnet}
\alias{my.cv.glmnet}
\title{Title}
\usage{
my.cv.glmnet(alpha, xdata, ydata, penalty.factor, nlambda = 1000,
  nfolds = 10, lambda.min.ratio = 0.001, seed = 1985, ...)
}
\arguments{
\item{show.messages}{}
}
\description{
Title
}
