% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettify.labels.R
\name{prettify.labels}
\alias{prettify.labels}
\title{Title}
\usage{
prettify.labels(label, ret.value = "string.short")
}
\arguments{
\item{label}{}
}
\description{
Title
}
\examples{
prettify.labels('orphan.glmnet.83')
prettify.labels('orphan.103', 'base.model')
prettify.labels('orphan.degree_new.83', 'base.model')
prettify.labels('orphan.glmnet.83', 'base.model')
prettify.labels(c('orphan.glmnet.83', 'degree_new.orphan.83'))
prettify.labels('orphan.glmnet.83', 'string')
prettify.labels('orphan.glmnet.83', 'string.short')
prettify.labels('orphan.glmnet.83', 'string.base')
}
