% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_variables_survival.R
\name{test.all.clinical}
\alias{test.all.clinical}
\title{Test all survival variables that are continuous or binary}
\usage{
test.all.clinical(clinical, ydata)
}
\arguments{
\item{clinical}{clinical dataset from tcga.data package}

\item{ydata}{table with time and status variables}
}
\description{
Test all survival variables that are continuous or binary
}
