% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.glmsparsenet_custom.R
\name{my.cv.glmnet}
\alias{my.cv.glmnet}
\title{Title}
\usage{
my.cv.glmnet(prefix, alpha, xdata, ydata, penalty.factor, xdata.digest = NULL,
  nfolds = 10, lambda.min.ratio = 0.001, n.cores = 1, seed = 1985)
}
\arguments{
\item{show.messages}{}
}
\description{
Title
}
